Refactor based on CR comments
